+++ tests/integration/precompiles/bech32/test_methods.go
@@ -61,7 +61,7 @@ func (s *PrecompileTestSuite) TestHexToBech32() {
 			func() []interface{} {
 				return []interface{}{
 					s.keyring.GetAddr(0),
-					"cosmos",
+					sdk.GetConfig().GetBech32AccountAddrPrefix(),
 				}
 			},
 			func(data []byte) {
@@ -107,7 +107,7 @@ func (s *PrecompileTestSuite) TestBech32ToHex() {
 		malleate    func() []interface{}
 		postCheck   func(data []byte)
 		expError    bool
-		errContains string
+		errContains func() string
 	}{
 		{
 			"fail - invalid args length",
@@ -116,7 +116,9 @@ func (s *PrecompileTestSuite) TestBech32ToHex() {
 			},
 			func([]byte) {},
 			true,
-			fmt.Sprintf(cmn.ErrInvalidNumberOfArgs, 1, 0),
+			func() string {
+				return fmt.Sprintf(cmn.ErrInvalidNumberOfArgs, 1, 0)
+			},
 		},
 		{
 			"fail - empty bech32 address",
@@ -127,7 +129,9 @@ func (s *PrecompileTestSuite) TestBech32ToHex() {
 			},
 			func([]byte) {},
 			true,
-			"invalid bech32 address",
+			func() string {
+				return "invalid bech32 address"
+			},
 		},
 		{
 			"fail - invalid bech32 address",
@@ -138,7 +142,9 @@ func (s *PrecompileTestSuite) TestBech32ToHex() {
 			},
 			func([]byte) {},
 			true,
-			fmt.Sprintf("invalid bech32 address: %s", "cosmos"),
+			func() string {
+				return fmt.Sprintf("invalid bech32 address: %s", "cosmos")
+			},
 		},
 		{
 			"fail - decoding bech32 failed",
@@ -149,7 +155,9 @@ func (s *PrecompileTestSuite) TestBech32ToHex() {
 			},
 			func([]byte) {},
 			true,
-			"decoding bech32 failed",
+			func() string {
+				return "decoding bech32 failed"
+			},
 		},
 		{
 			"fail - invalid address format",
@@ -160,7 +168,13 @@ func (s *PrecompileTestSuite) TestBech32ToHex() {
 			},
 			func([]byte) {},
 			true,
-			"address max length is 255",
+			func() string {
+				if addrVerifier := sdk.GetConfig().GetAddressVerifier(); addrVerifier != nil {
+					err := addrVerifier(sdk.AccAddress(make([]byte, 256)))
+					return err.Error()
+				}
+				return "address max length is 255"
+			},
 		},
 		{
 			"success - valid bech32 address",
@@ -178,7 +192,9 @@ func (s *PrecompileTestSuite) TestBech32ToHex() {
 				s.Require().Equal(s.keyring.GetAddr(0), addr)
 			},
 			false,
-			"",
+			func() string {
+				return ""
+			},
 		},
 	}
 
@@ -190,7 +206,7 @@ func (s *PrecompileTestSuite) TestBech32ToHex() {
 
 			if tc.expError {
 				s.Require().Error(err)
-				s.Require().ErrorContains(err, tc.errContains)
+				s.Require().ErrorContains(err, tc.errContains())
 				s.Require().Empty(bz)
 			} else {
 				s.Require().NoError(err)