+++ tests/integration/x/precisebank/test_genesis.go
@@ -47,6 +47,29 @@ func (s *GenesisTestSuite) SetupTestWithChainID(chainID testconstants.ChainID) {
 	})
 }
 
+func (s *GenesisTestSuite) adjustModuleBalance(expectedAmt sdkmath.Int) {
+	moduleAddr := s.network.App.GetAccountKeeper().GetModuleAddress(types.ModuleName)
+	balance := s.network.App.GetBankKeeper().GetBalance(s.network.GetContext(), moduleAddr, types.IntegerCoinDenom())
+
+	if balance.Amount.GT(expectedAmt) {
+		// Burn excess
+		err := s.network.App.GetBankKeeper().BurnCoins(
+			s.network.GetContext(),
+			types.ModuleName,
+			sdk.NewCoins(sdk.NewCoin(types.IntegerCoinDenom(), balance.Amount.Sub(expectedAmt))),
+		)
+		s.Require().NoError(err)
+	} else if balance.Amount.LT(expectedAmt) {
+		// Mint deficit
+		err := s.network.App.GetBankKeeper().MintCoins(
+			s.network.GetContext(),
+			types.ModuleName,
+			sdk.NewCoins(sdk.NewCoin(types.IntegerCoinDenom(), expectedAmt.Sub(balance.Amount))),
+		)
+		s.Require().NoError(err)
+	}
+}
+
 func (s *GenesisTestSuite) TestInitGenesis() {
 	tests := []struct {
 		name         string
@@ -71,12 +94,7 @@ func (s *GenesisTestSuite) TestInitGenesis() {
 			func() {
 				// The network setup creates an initial balance of 1, so we need to mint 1 more
 				// to get to the expected amount of 2 for this test case
-				err := s.network.App.GetBankKeeper().MintCoins(
-					s.network.GetContext(),
-					types.ModuleName,
-					sdk.NewCoins(sdk.NewCoin(types.IntegerCoinDenom(), sdkmath.NewInt(1))),
-				)
-				s.Require().NoError(err)
+				s.adjustModuleBalance(sdkmath.NewInt(2))
 			},
 			types.NewGenesisState(
 				types.FractionalBalances{
@@ -106,12 +124,7 @@ func (s *GenesisTestSuite) TestInitGenesis() {
 			func() {
 				// The network setup creates an initial balance of 1, so we need to burn that
 				// to get to 0 balance for this test case
-				err := s.network.App.GetBankKeeper().BurnCoins(
-					s.network.GetContext(),
-					types.ModuleName,
-					sdk.NewCoins(sdk.NewCoin(types.IntegerCoinDenom(), sdkmath.NewInt(1))),
-				)
-				s.Require().NoError(err)
+				s.adjustModuleBalance(sdkmath.ZeroInt())
 			},
 			types.NewGenesisState(
 				types.FractionalBalances{
@@ -129,12 +142,7 @@ func (s *GenesisTestSuite) TestInitGenesis() {
 			func() {
 				// The network setup creates an initial balance of 1, so we need to mint 99 more
 				// to get to 100 total balance for this test case
-				err := s.network.App.GetBankKeeper().MintCoins(
-					s.network.GetContext(),
-					types.ModuleName,
-					sdk.NewCoins(sdk.NewCoin(types.IntegerCoinDenom(), sdkmath.NewInt(99))),
-				)
-				s.Require().NoError(err)
+				s.adjustModuleBalance(sdkmath.NewInt(100))
 			},
 			types.NewGenesisState(
 				types.FractionalBalances{
@@ -234,19 +242,7 @@ func (s *GenesisTestSuite) TestExportGenesis() {
 			"balances, no remainder",
 			func() *types.GenesisState {
 				// Burn the initial balance created by network setup, then mint the expected amount
-				err := s.network.App.GetBankKeeper().BurnCoins(
-					s.network.GetContext(),
-					types.ModuleName,
-					sdk.NewCoins(sdk.NewCoin(types.IntegerCoinDenom(), sdkmath.NewInt(1))),
-				)
-				s.Require().NoError(err)
-
-				err = s.network.App.GetBankKeeper().MintCoins(
-					s.network.GetContext(),
-					types.ModuleName,
-					sdk.NewCoins(sdk.NewCoin(types.IntegerCoinDenom(), sdkmath.NewInt(1))),
-				)
-				s.Require().NoError(err)
+				s.adjustModuleBalance(sdkmath.NewInt(1))
 
 				return types.NewGenesisState(
 					types.FractionalBalances{
@@ -261,19 +257,7 @@ func (s *GenesisTestSuite) TestExportGenesis() {
 			"balances, remainder",
 			func() *types.GenesisState {
 				// Burn the initial balance created by network setup, then mint the expected amount
-				err := s.network.App.GetBankKeeper().BurnCoins(
-					s.network.GetContext(),
-					types.ModuleName,
-					sdk.NewCoins(sdk.NewCoin(types.IntegerCoinDenom(), sdkmath.NewInt(1))),
-				)
-				s.Require().NoError(err)
-
-				err = s.network.App.GetBankKeeper().MintCoins(
-					s.network.GetContext(),
-					types.ModuleName,
-					sdk.NewCoins(sdk.NewCoin(types.IntegerCoinDenom(), sdkmath.NewInt(1))),
-				)
-				s.Require().NoError(err)
+				s.adjustModuleBalance(sdkmath.NewInt(1))
 
 				return types.NewGenesisState(
 					types.FractionalBalances{