+++ tests/integration/x/vm/test_grpc_query.go
@@ -386,7 +386,7 @@ func (s *KeeperTestSuite) TestQueryTxLogs() {
 	expLogs := []*types.Log{}
 	txHash := common.BytesToHash([]byte("tx_hash"))
 	txIndex := uint(1)
-	logIndex := uint(1)
+	logIndex := uint(0)
 
 	testCases := []struct {
 		msg      string
@@ -428,7 +428,6 @@ func (s *KeeperTestSuite) TestQueryTxLogs() {
 			txCfg := statedb.NewTxConfig(
 				txHash,
 				txIndex,
-				logIndex,
 			)
 			vmdb := statedb.New(
 				s.Network.GetContext(),
@@ -450,6 +449,8 @@ func (s *KeeperTestSuite) TestQueryParams() {
 	expParams := types.DefaultParams()
 	expParams.ActiveStaticPrecompiles = types.AvailableStaticPrecompiles
 	expParams.ExtraEIPs = nil
+	expParams.EvmDenom = testconstants.ExampleAttoDenom
+	expParams.ExtendedDenomOptions = &types.ExtendedDenomOptions{ExtendedDenom: testconstants.ExampleAttoDenom}
 
 	res, err := s.Network.GetEvmClient().Params(ctx, &types.QueryParamsRequest{})
 	s.Require().NoError(err)
@@ -573,7 +574,7 @@ func (s *KeeperTestSuite) TestEstimateGas() {
 				return types.TransactionArgs{}
 			},
 			true,
-			ethparams.TxGasContractCreation,
+			53793,
 			false,
 			config.DefaultGasCap,
 		},
@@ -698,7 +699,7 @@ func (s *KeeperTestSuite) TestEstimateGas() {
 				}
 			},
 			true,
-			1187108,
+			1197697,
 			false,
 			config.DefaultGasCap,
 		},
@@ -726,7 +727,7 @@ func (s *KeeperTestSuite) TestEstimateGas() {
 				}
 			},
 			true,
-			51880,
+			52669,
 			false,
 			config.DefaultGasCap,
 		},
@@ -808,7 +809,7 @@ func (s *KeeperTestSuite) TestEstimateGas() {
 				}
 			},
 			true,
-			1187108,
+			1197697,
 			true,
 			config.DefaultGasCap,
 		},
@@ -837,7 +838,7 @@ func (s *KeeperTestSuite) TestEstimateGas() {
 				}
 			},
 			true,
-			51880,
+			52669,
 			true,
 			config.DefaultGasCap,
 		},
@@ -1066,7 +1067,7 @@ func (s *KeeperTestSuite) TestEstimateGasWithStateOverrides() {
 				return string(bz)
 			},
 			true,
-			49140,
+			52114,
 			false,
 			config.DefaultGasCap,
 		},