+++ x/ibc/transfer/keeper/keeper.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/cosmos/ibc-go/v10/modules/apps/transfer/keeper"
 	transfertypes "github.com/cosmos/ibc-go/v10/modules/apps/transfer/types"
-	porttypes "github.com/cosmos/ibc-go/v10/modules/core/05-port/types"
 
+	"cosmossdk.io/core/address"
 	corestore "cosmossdk.io/core/store"
 
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -25,9 +25,9 @@ type Keeper struct {
 // NewKeeper creates a new IBC transfer Keeper instance
 func NewKeeper(
 	cdc codec.BinaryCodec,
+	addressCodec address.Codec,
 	storeService corestore.KVStoreService,
 
-	ics4Wrapper porttypes.ICS4Wrapper,
 	channelKeeper transfertypes.ChannelKeeper,
 	msgRouter transfertypes.MessageRouter,
 	authKeeper types.AccountKeeper,
@@ -37,13 +37,13 @@ func NewKeeper(
 ) Keeper {
 	// create the original IBC transfer keeper for embedding
 	transferKeeper := keeper.NewKeeper(
-		cdc, storeService, nil,
-		ics4Wrapper, channelKeeper, msgRouter,
+		cdc, addressCodec, storeService,
+		channelKeeper, msgRouter,
 		authKeeper, bankKeeper, authority,
 	)
 
 	return Keeper{
-		Keeper:        &transferKeeper,
+		Keeper:        transferKeeper,
 		bankKeeper:    bankKeeper,
 		erc20Keeper:   erc20Keeper,
 		accountKeeper: authKeeper,