+++ x/vm/keeper/keeper.go
@@ -38,11 +38,11 @@ type Keeper struct {
 	// - storing Bloom filters by block height. Needed for the Web3 API.
 	storeKey storetypes.StoreKey
 
-	// key to access the transient store, which is reset on every block during Commit
-	transientKey storetypes.StoreKey
+	// key to access the object store, which is reset on every block during Commit
+	objectKey storetypes.StoreKey
 
-	// KVStore Keys for modules wired to app
-	storeKeys map[string]*storetypes.KVStoreKey
+	// Store Keys for modules wired to app
+	storeKeys map[string]storetypes.StoreKey
 
 	// the address capable of executing a MsgUpdateParams message. Typically, this should be the x/gov module account.
 	authority sdk.AccAddress
@@ -80,13 +80,21 @@ type Keeper struct {
 	// evmMempool is the custom EVM appside mempool
 	// if it is nil, the default comet mempool will be used
 	evmMempool *evmmempool.ExperimentalEVMMempool
+
+	// virtualFeeCollection enabling will use "Virtual" methods from the bank module to accumulate
+	// fees to the fee collector module in the endBlocker instead of using regular sends during tx execution.
+	virtualFeeCollection bool
+
+	// defaultEvmCoinInfo is the default EVM coin info used when evmCoinInfo is not initialized in the state,
+	// mainly for historical queries.
+	defaultEvmCoinInfo types.EvmCoinInfo
 }
 
 // NewKeeper generates new evm module keeper
 func NewKeeper(
 	cdc codec.BinaryCodec,
-	storeKey, transientKey storetypes.StoreKey,
-	keys map[string]*storetypes.KVStoreKey,
+	storeKey, objectKey storetypes.StoreKey,
+	keys []storetypes.StoreKey,
 	authority sdk.AccAddress,
 	ak types.AccountKeeper,
 	bankKeeper types.BankKeeper,
@@ -109,6 +117,10 @@ func NewKeeper(
 
 	bankWrapper := wrappers.NewBankWrapper(bankKeeper)
 	feeMarketWrapper := wrappers.NewFeeMarketWrapper(fmk)
+	storeKeys := make(map[string]storetypes.StoreKey)
+	for _, k := range keys {
+		storeKeys[k.Name()] = k
+	}
 
 	// set global chain config
 	ethCfg := types.DefaultChainConfig(evmChainID)
@@ -125,17 +137,31 @@ func NewKeeper(
 		stakingKeeper:    sk,
 		feeMarketWrapper: feeMarketWrapper,
 		storeKey:         storeKey,
-		transientKey:     transientKey,
+		objectKey:        objectKey,
 		tracer:           tracer,
 		consensusKeeper:  consensusKeeper,
 		erc20Keeper:      erc20Keeper,
-		storeKeys:        keys,
+		storeKeys:        storeKeys,
 	}
+}
+
+// EnableVirtualFeeCollection switches fee deduction for evm transactions to use the virtual fee collection of the
+// bank keeper via the object store.
+// Note: Do NOT use this if your chain does not have an 18 decimal point precision gas token.
+func (k *Keeper) EnableVirtualFeeCollection() {
+	k.virtualFeeCollection = true
 }
 
 // Logger returns a module-specific logger.
 func (k Keeper) Logger(ctx sdk.Context) log.Logger {
 	return ctx.Logger().With("module", types.ModuleName)
+}
+
+// WithDefaultEvmCoinInfo set default EvmCoinInfo
+func (k *Keeper) WithDefaultEvmCoinInfo(coinInfo types.EvmCoinInfo) *Keeper {
+	k.defaultEvmCoinInfo = coinInfo
+	types.SetDefaultEvmCoinInfo(coinInfo)
+	return k
 }
 
 // ----------------------------------------------------------------------------
@@ -144,11 +170,11 @@ func (k Keeper) Logger(ctx sdk.Context) log.Logger {
 // ----------------------------------------------------------------------------
 
 // EmitBlockBloomEvent emit block bloom events
-func (k Keeper) EmitBlockBloomEvent(ctx sdk.Context, bloom ethtypes.Bloom) {
+func (k Keeper) EmitBlockBloomEvent(ctx sdk.Context, bloom []byte) {
 	ctx.EventManager().EmitEvent(
 		sdk.NewEvent(
 			types.EventTypeBlockBloom,
-			sdk.NewAttribute(types.AttributeKeyEthereumBloom, string(bloom.Bytes())),
+			sdk.NewAttribute(types.AttributeKeyEthereumBloom, string(bloom)),
 		),
 	)
 }
@@ -158,40 +184,26 @@ func (k Keeper) GetAuthority() sdk.AccAddress {
 	return k.authority
 }
 
-// GetBlockBloomTransient returns bloom bytes for the current block height
-func (k Keeper) GetBlockBloomTransient(ctx sdk.Context) *big.Int {
-	store := prefix.NewStore(ctx.TransientStore(k.transientKey), types.KeyPrefixTransientBloom)
-	heightBz := sdk.Uint64ToBigEndian(uint64(ctx.BlockHeight())) //nolint:gosec // G115 // won't exceed uint64
-	bz := store.Get(heightBz)
-	if len(bz) == 0 {
-		return big.NewInt(0)
+// CollectTxBloom collects all tx blooms and emit a single block bloom event
+func (k Keeper) CollectTxBloom(ctx sdk.Context) {
+	store := prefix.NewObjStore(ctx.ObjectStore(k.objectKey), types.KeyPrefixObjectBloom)
+	it := store.Iterator(nil, nil)
+	defer it.Close()
+
+	bloom := new(big.Int)
+	for ; it.Valid(); it.Next() {
+		bloom.Or(bloom, it.Value().(*big.Int))
+		store.Delete(it.Key())
 	}
 
-	return new(big.Int).SetBytes(bz)
+	k.EmitBlockBloomEvent(ctx, bloom.Bytes())
 }
 
-// SetBlockBloomTransient sets the given bloom bytes to the transient store. This value is reset on
+// SetTxBloom sets the given bloom bytes to the object store. This value is reset on
 // every block.
-func (k Keeper) SetBlockBloomTransient(ctx sdk.Context, bloom *big.Int) {
-	store := prefix.NewStore(ctx.TransientStore(k.transientKey), types.KeyPrefixTransientBloom)
-	heightBz := sdk.Uint64ToBigEndian(uint64(ctx.BlockHeight())) //nolint:gosec // G115 // won't exceed uint64
-	store.Set(heightBz, bloom.Bytes())
-}
-
-// ----------------------------------------------------------------------------
-// Tx
-// ----------------------------------------------------------------------------
-
-// SetTxIndexTransient set the index of processing transaction
-func (k Keeper) SetTxIndexTransient(ctx sdk.Context, index uint64) {
-	store := ctx.TransientStore(k.transientKey)
-	store.Set(types.KeyPrefixTransientTxIndex, sdk.Uint64ToBigEndian(index))
-}
-
-// GetTxIndexTransient returns EVM transaction index on the current block.
-func (k Keeper) GetTxIndexTransient(ctx sdk.Context) uint64 {
-	store := ctx.TransientStore(k.transientKey)
-	return sdk.BigEndianToUint64(store.Get(types.KeyPrefixTransientTxIndex))
+func (k Keeper) SetTxBloom(ctx sdk.Context, bloom *big.Int) {
+	store := ctx.ObjectStore(k.objectKey)
+	store.Set(types.ObjectBloomKey(ctx.TxIndex(), ctx.MsgIndex()), bloom)
 }
 
 // ----------------------------------------------------------------------------
@@ -226,23 +238,6 @@ func (k *Keeper) PostTxProcessing(
 // HasHooks returns true if hooks are set
 func (k *Keeper) HasHooks() bool {
 	return k.hooks != nil
-}
-
-// ----------------------------------------------------------------------------
-// Log
-// ----------------------------------------------------------------------------
-
-// GetLogSizeTransient returns EVM log index on the current block.
-func (k Keeper) GetLogSizeTransient(ctx sdk.Context) uint64 {
-	store := ctx.TransientStore(k.transientKey)
-	return sdk.BigEndianToUint64(store.Get(types.KeyPrefixTransientLogSize))
-}
-
-// SetLogSizeTransient fetches the current EVM log index from the transient store, increases its
-// value by one and then sets the new index back to the transient store.
-func (k Keeper) SetLogSizeTransient(ctx sdk.Context, logSize uint64) {
-	store := ctx.TransientStore(k.transientKey)
-	store.Set(types.KeyPrefixTransientLogSize, sdk.Uint64ToBigEndian(logSize))
 }
 
 // ----------------------------------------------------------------------------
@@ -366,23 +361,33 @@ func (k Keeper) GetMinGasPrice(ctx sdk.Context) math.LegacyDec {
 	return k.feeMarketWrapper.GetParams(ctx).MinGasPrice
 }
 
-// ResetTransientGasUsed reset gas used to prepare for execution of current cosmos tx, called in ante handler.
+// ResetTransientGasUsed reset gas used to prepare for execution of cosmos tx, called in EndBlocker.
 func (k Keeper) ResetTransientGasUsed(ctx sdk.Context) {
-	store := ctx.TransientStore(k.transientKey)
-	store.Delete(types.KeyPrefixTransientGasUsed)
+	store := prefix.NewObjStore(ctx.ObjectStore(k.objectKey),
+		types.KeyPrefixObjectGasUsed)
+	it := store.Iterator(nil, nil)
+
+	defer it.Close()
+
+	for ; it.Valid(); it.Next() {
+		store.Delete(it.Key())
+	}
 }
 
 // GetTransientGasUsed returns the gas used by current cosmos tx.
 func (k Keeper) GetTransientGasUsed(ctx sdk.Context) uint64 {
-	store := ctx.TransientStore(k.transientKey)
-	return sdk.BigEndianToUint64(store.Get(types.KeyPrefixTransientGasUsed))
+	store := ctx.ObjectStore(k.objectKey)
+	v := store.Get(types.ObjectGasUsedKey(ctx.TxIndex()))
+	if v == nil {
+		return 0
+	}
+	return v.(uint64)
 }
 
 // SetTransientGasUsed sets the gas used by current cosmos tx.
 func (k Keeper) SetTransientGasUsed(ctx sdk.Context, gasUsed uint64) {
-	store := ctx.TransientStore(k.transientKey)
-	bz := sdk.Uint64ToBigEndian(gasUsed)
-	store.Set(types.KeyPrefixTransientGasUsed, bz)
+	store := ctx.ObjectStore(k.objectKey)
+	store.Set(types.ObjectGasUsedKey(ctx.TxIndex()), gasUsed)
 }
 
 // AddTransientGasUsed accumulate gas used by each eth msgs included in current cosmos tx.
@@ -396,7 +401,7 @@ func (k Keeper) AddTransientGasUsed(ctx sdk.Context, gasUsed uint64) (uint64, er
 }
 
 // KVStoreKeys returns KVStore keys injected to keeper
-func (k Keeper) KVStoreKeys() map[string]*storetypes.KVStoreKey {
+func (k Keeper) KVStoreKeys() map[string]storetypes.StoreKey {
 	return k.storeKeys
 }