+++ x/vm/types/msg.go
@@ -356,5 +356,5 @@ func (m *MsgUpdateParams) ValidateBasic() error {
 
 // GetSignBytes implements the LegacyMsg interface.
 func (m MsgUpdateParams) GetSignBytes() []byte {
-	return sdk.MustSortJSON(AminoCdc.MustMarshalJSON(&m))
+	return AminoCdc.MustMarshalJSON(&m)
 }